refactor: delete script, move test governanceConfig
